# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamSync Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, reseller, client]
        - in: query
          name: service
          schema:
            type: string
            enum: [plex, jellyfin, emby]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: resellerId
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, reseller, client]
        service:
          type: string
          enum: [plex, jellyfin, emby]
          nullable: true
        serverId:
          type: integer
          nullable: true
        packageId:
          type: integer
          nullable: true
        active:
          type: boolean
        expiresAt:
          type: string
          nullable: true
        resellerId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, role, active, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, reseller, client]
        service:
          type: string
          enum: [plex, jellyfin, emby]
        serverId:
          type: integer
        packageId:
          type: integer
        expiresAt:
          type: string
        resellerId:
          type: integer
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, reseller, client]
        service:
          type: string
          enum: [plex, jellyfin, emby]
        serverId:
          type: integer
        packageId:
          type: integer
        active:
          type: boolean
        expiresAt:
          type: string
        resellerId:
          type: integer
...